Split build file read effect tests
